break up components, add commands
