+++ src/install/internal/command/templates/Caddyfile.tmpl
@@ -1,0 +1,8 @@
+{{ .Hostname }} {
+    reverse_proxy localhost:3000
+}
+
+ws.{{ .Hostname }} {
+    reverse_proxy localhost:8080
+}
+
